# docs/swagger.yaml
definitions:
...
  dto.DeleteCommentResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.DeleteUserResponse:
    properties:
      message:
        type: string
    type: object
  dto.GetAllCommentResponse:
    properties:
      Photo:
        $ref: '#/definitions/dto.PhotoDataForComment'
      User:
        $ref: '#/definitions/dto.UserDataForComment'
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  dto.LoginUserRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  dto.LoginUserResponse:
    properties:
      token:
        type: string
    type: object
  dto.NewCommentRequest:
    properties:
      message:
        type: string
      photo_id:
        type: integer
    required:
    - message
    - photo_id
    type: object
  dto.NewCommentResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  dto.NewUserRequest:
    properties:
      age:
        type: integer
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  dto.NewUserResponse:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  dto.PhotoDataForComment:
    properties:
      caption:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      user_id:
        type: integer
    type: object
...
  dto.UpdateCommentRequest:
    properties:
      message:
        type: string
    type: object
  dto.UpdateCommentResponse:
    properties:
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  dto.UpdateUserRequest:
    properties:
      email:
        type: string
      username:
        type: string
    type: object
  dto.UpdateUserResponse:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  dto.UserDataForComment:
    properties:
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
...
  errs.MessageErrData:
    properties:
      error:
        example: BAD_REQUEST
        type: string
      message:
        example: This is an error message
        type: string
      statusCode:
        example: 400
        type: integer
    type: object
info:
  contact: {}
paths:
  /comment:
    get:
      description: View all comment by json
      parameters:
      - default: Bearer <Add your access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetAllCommentResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errs.MessageErrData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.MessageErrData'
      summary: View all comment
      tags:
      - comment
    post:
      consumes:
      - application/json
      description: Create a comment by json
      parameters:
      - description: Create a comment request body
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/dto.NewCommentRequest'
      - default: Bearer <Add your access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.NewCommentResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errs.MessageErrData'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.MessageErrData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.MessageErrData'
      summary: Create a comment
      tags:
      - comment
  /comment/{commentId}:
    delete:
      description: Delete a comment by param
      parameters:
      - description: comment ID request
        in: path
        name: commentId
        required: true
        type: integer
      - default: Bearer <Add your access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteCommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.MessageErrData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errs.MessageErrData'
      summary: Delete a comment
      tags:
      - comment
    put:
      consumes:
      - application/json
      description: Update a Comment by json
      parameters:
      - description: Update a comment request body
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCommentRequest'
      - description: comment ID request
        in: path
        name: commentId
        required: true
        type: integer
      - default: Bearer <Add your access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateCommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.MessageErrData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errs.MessageErrData'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.MessageErrData'
      summary: Update a Comment
      tags:
      - comment
  /comment/photo/{photoId}:
    get:
      consumes:
      - application/json
      description: View all comments of a photo by json
      parameters:
      - description: photo  ID request
        in: path
        name: photoId
        required: true
        type: integer
      - default: Bearer <Add your access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetAllCommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.MessageErrData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errs.MessageErrData'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.MessageErrData'
      summary: View all comments of a photo
      tags:
      - comment
  /comment/user/{userId}:
    get:
      consumes:
      - application/json
      description: View all comments of a user by json
      parameters:
      - description: user ID request
        in: path
        name: userId
        required: true
        type: integer
      - default: Bearer <Add your access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetAllCommentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.MessageErrData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errs.MessageErrData'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.MessageErrData'
      summary: View all comments of a user
      tags:
      - comment
...
  /users:
    delete:
      description: Delete a user by JWT from header
      parameters:
      - default: Bearer <Add your access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errs.MessageErrData'
      summary: Delete a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update a user by json
      parameters:
      - description: Update a user request body
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      - default: Bearer <Add your access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.MessageErrData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errs.MessageErrData'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.MessageErrData'
      summary: Update a user
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Login by json
      parameters:
      - description: Login user request body
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.MessageErrData'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.MessageErrData'
      summary: Login
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Register a user by json
      parameters:
      - description: Create user request body
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.NewUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.NewUserResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.MessageErrData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.MessageErrData'
      summary: Register a user
      tags:
      - users
swagger: "2.0"